DROP: arm64: dts: tegra210: increase PWRGOOD and OSC_POSTPWR for tests.
From Chris Bareng:
A resume problem was caused by the PWRGOOD and OSC_POSTPWR values in register
APBDEV_PMC_PWRGOOD_TIMER_0 being too small compared to the updated
timeslot periods of 5.120ms. Since LDO7 is at timeslot 3, it turns on
15.36ms after the resume sequence begins. The driver had only 3.876ms
for PWRGOOD and 3.052ms for OSC_POSTPWR, so the total 6.928ms delay was
not enough and expired before LDO7 turned on at timeslot 3.

Ideally we should separate SOC and LDO7 as we previously had done, so
we've settled on using PWRGOOD of 6ms (programmed value of 0xC5 ->
69*30.51us = 6.010ms), with some extra margin for SOC's 5.12ms at
timeslot 1, and OSC_POSTPWR of 12ms (programmed value of 0x63 ->
99*4*30.51us = 12.082ms), for a PWRGOOD+OSC_POSTPWR total period target
of 18ms (actual 18.092ms) with margin for the 15.36ms needed for LDO7 at
timeslot 3.

BUG=chrome-os-partner:46608
TEST=boots on smaug.

Change-Id: I12d77f5ecec7acf66072139ded9a06e06ba38b88
Signed-off-by: Hung-Te Lin <hungte@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/308710