CHROMIUM: iio:light:ap3223 Add more delays during init
This adds delays at points of failure observed by QA during the initialization
of ALS. The testing scenario is explained in the bug.

BUG=chrome-os-partner:46838
BRANCH=none
TEST=Tested by removing and inserting the driver repeatedly without any delays
(this was the quickest method to reproduce what QA saw) and with the fix the
driver doesn't fail. Also took the script that QA used to take the device
through reboot cycle and tested the device overnight (for about 400 reboot
cycles) and found that the issue doesn't happen with the fix.

Change-Id: I0cf41de18db7626795ff58001ace96a736fff225
Signed-off-by: Suresh Rajashekara <sureshraj@google.com>
Reviewed-on: https://chromium-review.googlesource.com/308959
Commit-Ready: Suresh Rajashekara <sureshraj@chromium.org>
Tested-by: Suresh Rajashekara <sureshraj@chromium.org>
Reviewed-by: Bryan Freed <bfreed@google.com>