UPSTREAM: ASoC: Intel: Skylake: Fix bit depth when querying the NHLT blob
Bps calculation is not correct as this needs to be based on valid
bit depth.  16 bit fmt bit depth is 16 bit and for 24 and 32 bit
as it is container size This patch fixes the bps.

Signed-off-by: Jeeja KP <jeeja.kp@intel.com>
Signed-off-by: Vinod Koul <vinod.koul@intel.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
Link: git://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git
br: for-next
(cherry picked from commit 4dfdd449760654cdedda2f40c23d7f3cb864b02a)
Signed-off-by: Vedang Patel <vedang.patel@intel.com>

BUG=chrome-os-partner:41280
TEST=with the whole series applied, audio works on SKL.
TEST=there are no regressions observed on BSW.

Change-Id: I4b4db06986d23d62125881c57454ed1affe05e85
Reviewed-on: https://chromium-review.googlesource.com/315311
Reviewed-by: Bowgo Tsai <bowgotsai@chromium.org>
Tested-by: Duncan Laurie <dlaurie@chromium.org>
Commit-Queue: Duncan Laurie <dlaurie@chromium.org>