FROMLIST: usb: dwc2: optionally assert phy "full reset" when waking up
On the rk3288 USB host-only port (the one that's not the OTG-enabled
port) the PHY can get into a bad state when a wakeup is asserted (not
just a wakeup from full system suspend but also a wakeup from
autosuspend).

We can get the PHY out of its bad state by asserting its "port reset",
but unfortunately that seems to assert a reset onto the USB bus so it
could confuse things if we don't actually deenumerate / reenumerate the
device.

We can also get the PHY out of its bad state by fully resetting it using
the reset from the CRU (clock reset unit), which does a more full
reset.  The CRU-based reset appears to actually cause devices on the bus
to be removed and reinserted, which fixes the problem (albeit in a hacky
way).

It's unfortunate that we need to do a full re-enumeration of devices at
wakeup time, but this is better than alternative of letting the bus get
wedged.

Conflicts:
    Documentation/devicetree/bindings/usb/dwc2.txt
    drivers/usb/dwc2/core.h
    drivers/usb/dwc2/platform.c
...due to patches for leaving PHY on during suspend

BUG=chrome-os-partner:45838
TEST=see BUG

Change-Id: I24ab85ee0e52574470536217a2a8e93939c15695
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Signed-off-by: Yunzhi Li <lyz@rock-chips.com>
(am from https://patchwork.kernel.org/patch/7529101/)
Reviewed-on: https://chromium-review.googlesource.com/306450