FROMLIST: ARM: dts: rockchip: Point rk3288 dwc2 usb at the full PHY reset
The "host1" port (AKA the dwc2 port that isn't the OTG port) on rk3288
has a hardware errata that causes everything to get confused when we get
a remote wakeup.  We'll use the reset that's in the CRU to reset the
port when it's in a bad state.

Note that we add the reset to both dwc2 controllers even though only one
has the errata in case we find some other use for this reset that's
unrelated to the current hardware errata.  Only the host port gets the
quirk property, though.

BUG=chrome-os-partner:45838
TEST=see BUG

Change-Id: Ia83cea8fe3a5c5d5078947944a0de9713649200a
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/7529121/)
Reviewed-on: https://chromium-review.googlesource.com/307834