FROMLIST: usb: dwc2: bus suspend/resume that's not hibernate
This is an attempt to rehash commit 0cf884e819e0 ("usb: dwc2: add bus
suspend/resume for dwc2") on ToT.  That commit was reverted in commit
b0bb9bb6ce01 ("Revert "usb: dwc2: add bus suspend/resume for dwc2"")
because apparently it broke the Altera SOCFPGA.

With all the changes that have happened to dwc2 in the meantime, it's
possible that the Altera SOCFPGA will just magically work with this
change now.  ...and it would be good to get bus suspend/resume
implemented.

BUG=chrome-os-partner:46308, chromium:548336
TEST=w/ series both USB ports are happy; power_Resume is back to normal

Change-Id: I19f2e0d8ef6c6144bc4fd17237562ea37d7ba5b9
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(am from https://patchwork.kernel.org/patch/7529161/)
Reviewed-on: https://chromium-review.googlesource.com/306265