CHROMIUM: ARM: Add alt-syscall support
Add ARM support for changing a thread's syscall table at runtime using
the alt-syscall infrastructure.

Note that alt-syscall is not supported in kernel configurations using
OABI or EABI with OABI_COMPAT.  It's possible to support OABI, but it
requires a bit of re-work to the common alt-syscall infrastructure and
to the syscall entry path.  This should not be an issue since OABI is
not supported on ChromeOS or Android.

BUG=b:25408860
TEST=Build and boot on Minnie; syscalls still work with ALT_SYSCALL on.

Change-Id: If87c348262e493d8a422de76148e1613b24dce2e
Signed-off-by: Andrew Bresticker <abrestic@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/311003