UPSTREAM: ASoC: rockchip: i2s: change bclk and lrck according to sample rates
This patch sets the dividers autonomously.

when i2s works on master mode, and sample rates changed. We need to change
bclk and lrck at the same time for cpu internal side.

As the input source clock to the module is MCLK_I2S,
and by the divider of the module, the clock generator generates
SCLK and LRCK to transmitter and receiver.

BUG=chrome-os-partner:46934
TEST=Manually test with the commands
     arecord -D hw:0 -f S16_LE -c 2 --rate [8000-96000] -d 5 /tmp/foobar.wav
and
     aplay -D hw:0 /tmp/foobar.wav

Change-Id: Ie30e2b914c2b5c18125ee71e36ff443caa9c5c0e
Signed-off-by: Caesar Wang <wxt@rock-chips.com>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry-picked from the asoc tree at git://git.kernel.org/pub/scm/linux/kernel/git/broonie/sound.git
commit daf7fdaa9909c22a30504f3fbf64861f6fb222af)
Reviewed-on: https://chromium-review.googlesource.com/309865
Reviewed-by: Cheng-Yi Chiang <cychiang@chromium.org>