CHROMIUM: iwl7000: mvm: don't overwrite the key indices in D3 entry
When entering D3, we need to use hardcoded key indices because the
firmware requires that.  To do so, we are overwriting the HW key index
in the keyconf structure, which makes it impossible to reuse the
indices that were used before entering D3.  Additionally, we overwrite
all the non-PTK keys with index 1, because the firmware only allows
one non-PTK key to be set.  This is bad, because when we resume, we
may try to set more than one key with index 1, which will obviously
fail.

To fix this, allow the callers to set a pre-defined index to use in
iwl_mvm_set_sta_key() instead of relying on the hw_key_idx value from
the keyconf struct (which requires overwriting it).  In normal cases,
the caller can pass STA_KEY_IDX_INVALID, which will cause a new key
offset to be chosen.  During HW_RESTART, we pass the offset that is in
use.  And during D3 entry, we pass the hardcoded indices we need to
use.

Additionally, don't clear the fw_key_table in D3 entry, so that the
flags are still set with the pre-D3 values when exiting D3.

BUG=chrome-os-partner:46241
TEST=suspend with WoWLAN enabled, resume and disconnect; then check
that there are no "failed to remove key" messages in the kernel logs

Original-Change-Id: I954e8d46478aa1b676b58a0dd7bb1d5af4a537d4
Signed-off-by: Luca Coelho <luciano.coelho@intel.com>
Reviewed-on: https://chromium-review.googlesource.com/312461
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Samuel Tan <samueltan@chromium.org>
(cherry picked from commit dafa584d6ba4348f7eb39181b67516fb654cb068)

Change-Id: Ic464500cfa7f4df1f0356b2c5eeb78bb7dda8406
Reviewed-on: https://chromium-review.googlesource.com/313809
Reviewed-by: Sameer Nanda <snanda@chromium.org>
Commit-Queue: Sameer Nanda <snanda@chromium.org>
Tested-by: Sameer Nanda <snanda@chromium.org>