UPSTREAM: usb: dwc2: Return errors from PHY
When searching for PHYs, any error was treated as if the PHY did not
exist or was not specified. Thus the probe function did not
correctly return error conditions such as -EPROBE_DEFER.

Fixed so that only a non-existing PHY is ignored and any other error
is returned.

BUG=chrome-os-partner:46308
TEST=w/ series both USB ports are happy

Change-Id: I6f50bd46228bbcf9a5ef74d966058bf7109419c3
Acked-by: Eric Anholt <eric@anholt.net>
Reported-by: Alexander Aring <alex.aring@gmail.com>
Signed-off-by: John Youn <johnyoun@synopsys.com>
Signed-off-by: Stefan Wahren <stefan.wahren@i2se.com>
Tested-by: Marek Szyprowski <m.szyprowski@samsung.com>
Signed-off-by: Felipe Balbi <balbi@ti.com>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from git.kernel.org balbi/usb.git testing/next
 commit ce04e0d2d6633998fdf6dc4fbbc5c44852c06ae5)
Reviewed-on: https://chromium-review.googlesource.com/314282