CHROMIUM: driver-core: fixup modprobe status reporting for async firmware test
While this module can uncover bugs in some cases, it is insufficient in
two aspects:

(1) It doesn't print a clear message to the log; if we manage to hit the
    "not found" case, it will confusingly print both "No firmware found"
    and "firmware found". Let's print only one of those messages.

(2) There are other symptoms of failure, even if we don't see a page
    fault / NULL dereference; for instance, the firmware core might see
    a junk or empty string for the file name, and fail to load the
    firmware. This is also a symptom of use-after-free.

For problem #2, it makes sense to actually wait for the async request to
complete, and verify the load status. If the firmware failed to load,
then the module init routine can report an error. If the firmware loaded
successfully, then we report success.

Note that a "success" still doesn't necessarily mean we have no
use-after-free bug; it could just mean that the firmware loader won the
race.

BUG=chromium:557355
TEST=build and load module

Change-Id: I5fba7e3d091dbe6ab7ddd51dcf04256833fa110f
Signed-off-by: Brian Norris <briannorris@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/313156
Reviewed-by: Douglas Anderson <dianders@chromium.org>