FIXUP: usb: dwc2: host: Fix missing device insertions
We landed commit 387624c784fc ("FROMLIST: usb: dwc2: host: Fix missing
device insertions") and then it got a few minor changes landing
upstream.  This catches us up to the version that's in Felipe Balbi's
testing branch.

---

If you've got your interrupt signals bouncing a bit as you insert your
USB device, you might end up in a state when the device is connected but
the driver doesn't know it.

Specifically, the observed order is:
 1. hardware sees connect
 2. hardware sees disconnect
 3. hardware sees connect
 4. dwc2_port_intr() - clears connect interrupt
 5. dwc2_handle_common_intr() - calls dwc2_hcd_disconnect()

Now you'll be stuck with the cable plugged in and no further interrupts
coming in but the driver will think we're disconnected.

We'll fix this by checking for the missing connect interrupt and
re-connecting after the disconnect is posted.  We don't skip the
disconnect because if there is a transitory disconnect we really want to
de-enumerate and re-enumerate.

Notes:
1. As part of this change we add a "force" parameter to
   dwc2_hcd_disconnect() so that when we're unloading the module we
   avoid the new behavior.  The need for this was pointed out by John
   Youn.
2. The bit of code needed at the end of dwc2_hcd_disconnect() is
   exactly the same bit of code from dwc2_port_intr().  To avoid
   duplication, we refactor that code out into a new function
   dwc2_hcd_connect().

BUG=chrome-os-partner:46308
TEST=w/ series both USB ports are happy

Change-Id: I4c9ced2d48d859d2fb08637d7b994bdebde22b73
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Acked-by: John Youn <johnyoun@synopsys.com>
Tested-by: John Youn <johnyoun@synopsys.com>
Signed-off-by: Felipe Balbi <balbi@ti.com>
(cherry picked from git.kernel.org balbi/usb.git testing/next
 commit 0addc634e5c00b02b62ff10e630fd782dbfbdebf)
Reviewed-on: https://chromium-review.googlesource.com/314287