UPSTREAM: ARM: dts: rockchip: make sure edp_24m is associated to xin24m on veyron
The edp-24m clock has two possible sources: the 24MHz oscillator as well
as an external 27MHz input. The power-on-default is the 27MHz clock which
is not supplied on all Rockchip boards. While on all current boards and
also all Veyron Chromebooks the bootloader seems to adapt the muxing to
the internal source, this doesn't seem to be the case on headless veyron
devices like brain and mickey making the edp-24m clock an orphan.
On the hardware side the 27m input also is not connected at all.

With the upcoming deferral of orphan-clocks this results in the power-
domain code deferring, as it cannot request the needed clock and if the
synchronous reset is sucessfullat all in this case is also unknown.

So fix that by making sure, the edp-24m clock is muxed to the internal
24MHz oscillator at all times.

BUG=None
TEST=/sys/kernel/debug/clk/clk_summary shows sclk_edp_24m correct now

Change-Id: I2fdfa4543937e79f2537cf409334cea8eedb770b
Signed-off-by: Heiko Stuebner <heiko.stuebner@collabora.com>
Signed-off-by: Douglas Anderson <dianders@chromium.org>
(cherry picked from git.kernel.org mmind/linux-rockchip.git
 linuxrockchip/v4.5-armsoc/dts32
 commit ba5810ab0c84069a73b4cd30582c77568158635f)
Reviewed-on: https://chromium-review.googlesource.com/316004
Reviewed-by: Brian Norris <briannorris@chromium.org>