CHROMIUM: drm: rockchip: If disabling DMC, do it before changing mode
dmc_notify() configures a VOP LINE_FLAG interrupt to fire when the last
line of pixels has been displayed on screen based on the current
vop->crtc.hwmode, and then synchronously waits for this interrupt.

During mode_set, the drm crtc helper first calls the
vop_crtc_mode_set(), and only updates vop->crtc.hwmode after it
returns.  However, it is vop_crtc_mode_set() that actually changes the
pixel clock.

Thus, there is a short window in time where vop->crtc.hwmode does not
match the actual hardware mode - it still has old mode parameters, but
the vop is actually scanning out using a new mode and clock.

If we call dmc_notify() during this time, it may configure a
LINE_FLAG interrupt that will never fire.

In fact, this is exactly what happens when we transition from a mode
where vblank *was* long enough to do DMC, to a mode where DMC must be
disabled.

Fix this by disabling DMC earlier in vop_crtc_mode_set(), before
changing the pixel clock, to ensure vop->crtc.hwmode still accurately
reflects the real hw mode.

Note that with old code, we ended up with a hang with:

- vop_crtc_mode_set()
- rockchip_dmc_disable()
- raw_notifier_call_chain()
- notifier_call_chain()
- rk3288_dmc_enable_notify()
- rk3288_dmcfreq_target()
- flush_work()
- wait_for_completion()

...and then that waits for:

- process_one_work()
- rk3288_dmcfreq_work()
- clk_set_rate()
- clk_change_rate()
- rk3288_dmcclk_set_rate()
- rockchip_dmc_wait()
- raw_notifier_call_chain()
- notifier_call_chain()
- dmc_notify()
- wait_for_completion()

If we move the disable to the top of the function everything becomes
reliable.  It's OK to leave the enable at the bottom of the function
because that's not blocking.

BUG=chrome-os-partner:48332
TEST=No more hangs, no more printouts about timeouts.

Change-Id: Ic78730293ba2d8d6c27ee833d27c519b59c6b717
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/316660
Reviewed-by: Derek Basehore <dbasehore@chromium.org>
(cherry picked from commit 318b71c013a83a6ac226f20f5a3505b74dac0378)
Reviewed-on: https://chromium-review.googlesource.com/317452