CHROMIUM: drm: rockchip: supress ERESTARSYS error message
The wait for completion in rockchip drm GEM acquire ioctl can abort when
there is an external interrupt, like when mouse is moved, it is not a problem
because syscall is then restarted from user space. This causes lots of
annoying kernel error messages. This change suppresses error message for
ERESTARTSYS error code.

BUG=chromium:569268
TEST=when running X server move mouse and examine kernel log

Change-Id: Idfa620d3e91437f5d2affdc5f5ab4038f3470d2b
Signed-off-by: Dominik Behr <dbehr@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/318835
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>
Reviewed-by: 智情 姚 <mark.yao@rock-chips.com>